ai-backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.3",
    "langchain-ollama>=1.0.1",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
]
